NMS-17739, initial commit

+++ opennms-dao-api/src/main/java/org/opennms/netmgt/dao/api/AlarmDao.java
@@ -68,5 +68,5 @@ public interface AlarmDao extends LegacyOnmsDao<OnmsAlarm, Integer> {
      */
     long getNumSituations();
 
-    int getNumAlarmsLastHours(int hours );
+    long getNumAlarmsLastHours(int hours );
 }
